Revert "Move reports controller into dashboards controller"
This reverts commit 9d0bb80ef5283a50af5b17361b0b72d250c9a44b.

The reports are really their own thing; the front page is effectively
the dashboard so we don't need to put them inside another dashboard
