+++ config/routes.rb
@@ -170,7 +170,7 @@ Rails.application.routes.draw do
   get '/catalog/parent/:parent_id/:id', to: 'catalog#show', as: :parent_solr_document
   get "/iiif/lookup/:prefix/:naan/:arkid", to: 'catalog#lookup_manifest', as: :lookup_manifest
 
-  get "/reports/identifiers_to_reconcile", to: "dashboard#identifiers_to_reconcile", as: :identifiers_to_reconcile
+  get "/reports/identifiers_to_reconcile", to: "reports#identifiers_to_reconcile", as: :identifiers_to_reconcile
 
   mount BrowseEverything::Engine => '/browse'
   mount Valhalla::Engine => '/'
